+++ services/users/__tests__/jest.config.db.ts
@@ -1,0 +1,9 @@
+import { Config } from 'jest';
+import defaultConfig from './jest.config.default';
+
+const config: Config = {
+  ...defaultConfig,
+  roots: ['./db'],
+};
+
+export default config;